requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
blessed==1.19.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.4
colorful==0.5.5
comm==0.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
dm-tree==0.1.7
entrypoints==0.4
etils==0.9.0
executing==1.2.0
filelock==3.8.2
flatbuffers==22.12.6
flwr==1.1.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
gpustat==1.0.0
grpcio==1.43.0
h5py==3.7.0
idna==3.4
importlib-resources==5.10.1
ipykernel==6.19.0
ipython==8.7.0
iterators==0.0.2
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
msgpack==1.0.4
multidict==6.0.3
nest-asyncio==1.5.6
numpy==1.23.5
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.0
opencensus-context==0.1.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
ray==2.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
smart-open==6.2.0
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io==0.29.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
